services/phame/requirements.txt
amqp==2.2.2
Babel==2.5.3
beautifulsoup4==4.4.0
billiard==3.5.0.3
celery==4.2.0rc4
coverage==4.5.2
click==6.7
flake8===3.6.0
Flask==1.0.2
flask-login
flask_debugtoolbar==0.10.1
flask_cors==3.0.7
Flask-JSGlue
flask_migrate==2.2.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
flask-wtf
gunicorn==19.9.0
itsdangerous==0.24
Jinja2==2.10.1
kombu==4.2.0
MarkupSafe==1.0
pandas==0.23.1
psutil==5.6.2
psycopg2==2.7.4
pytz==2018.3
redis==2.10.6
requests
tornado==5.0.2
vine==1.1.4
Werkzeug==0.15.3
requests
IPython
Werkzeug==0.15.3
